server/pyproject.toml
[project]
name = "leanaide_serv"
version = "0.1.0"
description = "Leanaide Streamlit Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit-sortables>=0.3.1",
    "watchdog>=6.0.0",
    "pymupdf>=1.26.1",
    "st-copy>=1.1.2",
    "openai>=1.91.0",
    "streamlit>=1.46.0",
    "python-dotenv>=1.1.1",
    "ollama>=0.5.1",
    "torch>=2.7.1",
    "transformers>=4.53.1",
    "huggingface-hub>=0.33.2",
    "timm>=1.0.16",
]
